Add searchContacts endpoint
Added SplType, a file from the PHP runtime to remove dependency on PHP
libraries that may not be included in install.

Modified supportFunctions to better handle the JSON formatting that is
sent out from the server

2 TODOs still exist on how to handle error handling for searchContacts
endpoint, but the API is ready in the event that no error message wishes
to be returned.